+++ front/scripts/mercury/utils/variantTesting.ts
@@ -36,5 +36,36 @@ module Mercury.Utils.VariantTesting {
 			optimizely.push(['trackEvent', eventName]);
 		}
 	}
+
+	/**
+	 * Integrates Optimizely with Universal Analytics
+	 *
+	 * @param {[]} dimensions
+	 * @returns {[]}
+	 */
+	export function integrateOptimizelyWithUA (dimensions: any[]): any[] {
+		var optimizely = window.optimizely,
+			experimentId: string,
+			dimension: number,
+			experimentName: string,
+			variationName: string;
+
+		if (optimizely && optimizely.activeExperiments) {
+			optimizely.activeExperiments.forEach((experimentId: string): void => {
+				if (
+					optimizely.allExperiments.hasOwnProperty(experimentId) &&
+					typeof optimizely.allExperiments[experimentId].universal_analytics === 'object'
+				) {
+					dimension = optimizely.allExperiments[experimentId].universal_analytics.slot;
+					experimentName = optimizely.allExperiments[experimentId].name;
+					variationName = optimizely.variationNamesMap[experimentId];
+
+					dimensions[dimension] = `Optimizely ${experimentName} (${experimentId}): ${variationName}`;
+				}
+			});
+		}
+
+		return dimensions;
+	}
 }