+++ server/facets/auth/join.ts
@@ -1,5 +1,6 @@
 /// <reference path='../../../typings/hapi/hapi.d.ts' />
 import authUtils = require('../../lib/AuthUtils');
+import caching = require('../../lib/Caching');
 
 interface JoinViewContext {
 	title: string;
@@ -13,9 +14,10 @@ interface JoinViewContext {
 	signupHref: string;
 }
 
-function get (request: Hapi.Request, reply: any): void {
+function get (request: Hapi.Request, reply: any): Hapi.Response {
 	var context: JoinViewContext,
-		redirectUrl: string = request.query.redirect || '/';
+		redirectUrl: string = request.query.redirect || '/',
+		response: Hapi.Response;
 
 	if (request.auth.isAuthenticated) {
 		return reply.redirect(redirectUrl);
@@ -33,13 +35,16 @@ function get (request: Hapi.Request, reply: any): void {
 		signupHref: authUtils.getSignupUrlFromRedirect(redirectUrl)
 	};
 
-	return reply.view(
+	response = reply.view(
 		'auth-landing-page',
 		context,
 		{
 			layout: 'auth'
 		}
 	);
+
+	caching.disableCache(response);
+	return response;
 }
 
 export = get;