+++ test/specs/front/scripts/main/views/MediaLightboxView.js
@@ -21,10 +21,10 @@ test('if changeMediaOnTap changes current media', function () {
 		nextMediaWasTapped = false,
 		prevMediaWasTapped = false;
 
-	mediaLightboxView.handleClick(tapEventRight);
+	mediaLightboxView.callClickHandler(tapEventRight);
 	equal(nextMediaWasTapped, true);
 
-	mediaLightboxView.handleClick(tapEventLeft);
+	mediaLightboxView.callClickHandler(tapEventLeft);
 	equal(prevMediaWasTapped, true);
 });
 
@@ -44,32 +44,4 @@ test('if isCurrentMediaType method recognizes media', function () {
 
 	isMediaTypeImage = mediaLightboxView.isCurrentMediaType('image');
 	equal(isMediaTypeImage, true);
-});
-
-test('if getScreenArea correctly returns screen areas', function () {
-	var mediaLightboxView = this.subject({
-			viewportSize: {
-				width: 480
-			}
-		}),
-		tapEventRight = {
-			clientX: 400
-		},
-		tapEventLeft = {
-			clientX: 80
-		},
-		tapEventCenter = {
-			clientX: 240
-		},
-		screenArea,
-		screenAreas = mediaLightboxView.get('screenAreas');
-
-	screenArea = mediaLightboxView.getScreenArea(tapEventLeft);
-	equal(screenArea, screenAreas.left);
-
-	screenArea = mediaLightboxView.getScreenArea(tapEventRight);
-	equal(screenArea, screenAreas.right);
-
-	screenArea = mediaLightboxView.getScreenArea(tapEventCenter);
-	equal(screenArea, screenAreas.center);
 });