+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+## release-71 (2015-05-27 12:18 UTC)
+* [CONCF-582](https://wikia-inc.atlassian.net/browse/CONCF-582) CONCF-582 | Tell Optimizely to use UA Dimensions
+* fix-curated-content-invalid-typing: Fix curated content invalid typing
+
+## release-70 (2015-05-26 18:13 UTC)
+* [DAT-2812](https://wikia-inc.atlassian.net/browse/DAT-2812) DAT-2812 style image caption inside portable infoboxes
+* Upgrade-to-Ember-1.12: Upgrade to ember 1.12
+* [CONCF-505](https://wikia-inc.atlassian.net/browse/CONCF-505) CONCF-505 | trending videos for curated content mobile main pages
+* [CONCF-551](https://wikia-inc.atlassian.net/browse/CONCF-551) CONCF-551 Treat top level of curated content as a regular section
+* [CONCF-550](https://wikia-inc.atlassian.net/browse/CONCF-550) CONCF-550 Show pre-footer ad when trending articles and/or videos are present
+* [CONCF-549](https://wikia-inc.atlassian.net/browse/CONCF-549) CONCF-549
+* [HG-668](https://wikia-inc.atlassian.net/browse/HG-668) Hg 668 wiki redirect
+* trending-articles: CONCF-504 Trending articles
+
+## release-69 (2015-05-25 15:21 UTC)
+* [DAT-2812](https://wikia-inc.atlassian.net/browse/DAT-2812) DAT-2812 style image caption inside portable infoboxes
+
+## release-68 (2015-05-21 18:47 UTC)
+* [SOC-826](https://wikia-inc.atlassian.net/browse/SOC-826) SOC-826 Disable browser caching for HTML documents
+* feature/HG-690-ga-special-wikis: Add check for GA special wiki
+
+## release-67 (2015-05-21 17:53 UTC)
+* [CONCF-552](https://wikia-inc.atlassian.net/browse/CONCF-552) CONCF-552 Categories as Curated Content sections
+* feature/HG-656-tracking: Feature/hg 656 tracking
+* fixing-forgot-password-link: Fixing forgot-password link
+* fix-wrap-result: Use the returned value from wrapResult method
+* tslint-call-signature: Add check for call-signature to tslint and fix existing signatures
+
 ## release-66 (2015-05-19 20:21 UTC)
 * feature/performance: Add baseline performance metrics to send to Weppy
 * [HG-618](https://wikia-inc.atlassian.net/browse/HG-618) HG-618 adding error checking